Release 2.0.0 - MVP

+++ src/main/java/com/ehb/connected/domain/impl/feedbacks/service/FeedbackServiceImpl.java
@@ -5,13 +5,16 @@ import com.ehb.connected.domain.impl.feedbacks.entities.Feedback;
 import com.ehb.connected.domain.impl.feedbacks.dto.FeedbackDto;
 import com.ehb.connected.domain.impl.feedbacks.mappers.FeedbackMapper;
 import com.ehb.connected.domain.impl.feedbacks.repositories.FeedbackRepository;
+import com.ehb.connected.domain.impl.notifications.helpers.UrlHelper;
+import com.ehb.connected.domain.impl.notifications.service.NotificationServiceImpl;
 import com.ehb.connected.domain.impl.projects.entities.Project;
 import com.ehb.connected.domain.impl.projects.repositories.ProjectRepository;
+import com.ehb.connected.domain.impl.users.entities.Role;
 import com.ehb.connected.domain.impl.users.entities.User;
 import com.ehb.connected.domain.impl.users.services.UserService;
-import jakarta.persistence.EntityNotFoundException;
+import com.ehb.connected.exceptions.EntityNotFoundException;
+import com.ehb.connected.exceptions.UserUnauthorizedException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
@@ -24,39 +27,18 @@ public class FeedbackServiceImpl implements FeedbackService {
     private final ProjectRepository projectRepository;
     private final FeedbackRepository feedbackRepository;
     private final UserService userService;
-
+    private final NotificationServiceImpl notificationService;
     private final FeedbackMapper feedbackMapper;
 
-    private Feedback getFeedbackAndCheckPermissions(Principal principal, Long projectId, Long feedbackId) {
-        // Check if project exists
-        projectRepository.findById(projectId)
-                .orElseThrow(() -> new EntityNotFoundException("Project not found"));
-
-        // Find the feedback
-        final Feedback feedback = feedbackRepository.findById(feedbackId)
-                .orElseThrow(() -> new EntityNotFoundException("Feedback not found"));
-
-        // Check if feedback belongs to project
-        if (!feedback.getProject().getId().equals(projectId)) {
-            throw new IllegalArgumentException("Feedback does not belong to the specified project");
-        }
-
-        // Ensure that the user can modify this feedback
-        final User currentUser = userService.getUserByEmail(principal.getName());
-        if (!feedback.getUser().getId().equals(currentUser.getId())) {
-            throw new AccessDeniedException("You do not have permission to modify this feedback");
-        }
-
-        return feedback;
-    }
-
     @Override
-    public FeedbackDto giveFeedback(Principal principal, Long id, FeedbackCreateDto feedbackDto) {
-        final Project project = projectRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Project not found"));
+    public FeedbackDto giveFeedback(Principal principal, Long projectId, FeedbackCreateDto feedbackDto) {
+        final Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new EntityNotFoundException(Project.class, projectId));
 
         final User user = userService.getUserByEmail(principal.getName());
-        System.out.println("comment: " + feedbackDto.getComment());
+        if (user.getRole() == Role.STUDENT) {
+            throw new UserUnauthorizedException(user.getId());
+        }
         final Feedback feedback = new Feedback();
         feedback.setComment(feedbackDto.getComment());
         feedback.setUser(user);
@@ -64,12 +46,37 @@ public class FeedbackServiceImpl implements FeedbackService {
 
         feedbackRepository.save(feedback);
 
+        // Check if receiver exists and send notification
+        if (project.getProductOwner() != null) {
+            String destinationUrl = UrlHelper.BuildCourseAssignmentUrl(
+                    UrlHelper.Sluggify(project.getAssignment().getCourse().getName()),
+                    UrlHelper.Sluggify(project.getAssignment().getName()),
+                    "projects/" + project.getId(),
+                    "feedback");
+
+            notificationService.createNotification(
+                    project.getProductOwner(),
+                    "feedback has been written for  " + project.getTitle(),
+                    destinationUrl
+            );
+        }
+
+
+
         return feedbackMapper.toDto(feedback);
     }
 
     @Override
-    public FeedbackDto updateFeedback(Principal principal, Long id, Long feedbackId, FeedbackCreateDto feedbackDto) {
-        final Feedback feedback = getFeedbackAndCheckPermissions(principal, id, feedbackId);
+    public FeedbackDto updateFeedback(Principal principal, Long feedbackId, FeedbackCreateDto feedbackDto) {
+        //Retrieve the feedback by id
+        final Feedback feedback = feedbackRepository.findById(feedbackId)
+                .orElseThrow(() -> new EntityNotFoundException(Feedback.class, feedbackId));
+
+        // Ensure that the current user is the owner of the feedback
+        final User currentUser = userService.getUserByEmail(principal.getName());
+        if (!feedback.getUser().getId().equals(currentUser.getId())) {
+            throw new UserUnauthorizedException(currentUser.getId());
+        }
 
         // Update comment
         feedback.setComment(feedbackDto.getComment());
@@ -77,18 +84,24 @@ public class FeedbackServiceImpl implements FeedbackService {
     }
 
     @Override
-    public void deleteFeedback(Principal principal, Long id, Long feedbackId) {
-        final Feedback feedback = getFeedbackAndCheckPermissions(principal, id, feedbackId);
+    public void deleteFeedback(Principal principal,Long feedbackId) {
+        // Retrieve the feedback by id
+        final Feedback feedback = feedbackRepository.findById(feedbackId)
+                .orElseThrow(() -> new EntityNotFoundException(Feedback.class, feedbackId));
 
+        final User currentUser = userService.getUserByEmail(principal.getName());
+        if(!feedback.getUser().getId().equals(currentUser.getId())) {
+            throw new UserUnauthorizedException(currentUser.getId());
+        }
         // Delete the feedback
         feedbackRepository.delete(feedback);
     }
 
     @Override
-    public List<FeedbackDto> getAllFeedbackForProject(Principal principal, Long id) {
+    public List<FeedbackDto> getAllFeedbackForProject(Principal principal, Long projectId) {
         // Check if project exists
-        final Project project = projectRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Project not found"));
+        final Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new EntityNotFoundException(Project.class, projectId));
 
         // Return all feedback related to the project
         return feedbackMapper.toDtoList(feedbackRepository.findAllByProjectOrderByCreatedAtDesc(project));
